[squash] Rename and move to storageclass/
